requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
langchain-chroma
openai
chromadb
tiktoken
pypdf
python-dotenv
rapidocr-onnxruntime
pdf2image
unstructured
pillow
pydantic
lxml
matplotlib
tiktoken
open_clip_torch
torch
anthropic
voyageai
cohere
elasticsearch
pandas
numpy
PyPDF2
cohere